Only consider rescaling at runtime enabled when DPI awareness mode fits
Activating rescaling at runtime requires a proper DPI awareness mode to
be set. Currently, if setting the DPI awareness mode fails, rescaling
may still be activated if the user requested to. With this change,
setting the rescaling mode of a Display ensures that the correct DPI
awareness for the UI thread is set and, in case an error occurs, the
rescaling mode is not changed. It also adapts some faulty constants and
provides according test cases for setting the rescaling behavior.