Handle local resource references in Edge#setText()
This contribution allows Edge browser for win32 to serve webpages using setText method where the set html may refer to local resources. about:blank doesn't allow referencing local resources and hence, this contribution uses the URL to a temporary file to navigate to and serve the html allowing references to local resources. The temporary URL is not exposed to the clients and the APIs mimic about:blank in all the places where this URL is used to keep the consistency.

contributes to #213