[GTK4] Migrate DirectoryDialog from GtkFileChooser to GtkFileDialog
The FileChooser has been marked as deprecated with GTK 4.10 with the
FileDialog being its replacement. Using this new API poses a challenge,
as it requires the usage of the AsyncReadyCallback mechanism in order to
respond to the closure of the dialog.

Once the dialog has been opened via gtk_file_dialog_select_folder(), it
is necessary to call gtk_file_dialog_select_folder_finish() from within
the callback method. This method also returns the selected folder, which
has to be cached until after the dialog has been closed.

Following native methods have been added to GTK4:
- gtk_file_dialog_new
- gtk_file_dialog_select_folder
- gtk_file_dialog_select_folder_finish
- gtk_file_dialog_set_initial_folder