Need to throw the exception message to give more information (#1429)
If running on a system that does not have the required glibc
version then important information is lost to tell the user
what is wrong. For example:

version 'GLIBC_2.34' not found (required by .../libswt-pi3-gtk-4966r6.so)

The fallback is tried if that fails then its exception is added
as suppressed by the original exception and the original is thrown.